requirements.txt
asgiref==3.5.2
Django==4.1.2
django-phonenumber-field==7.0.0
django-widget-tweaks==1.4.12
phonenumbers==8.12.56
Pillow==9.2.0
sqlparse==0.4.3
